@@ -1,50 +1,148 @@
+<script>
+import { RouterLink, RouterView } from 'vue-router'
+
+export default {
+  data() {
+    return {
+      isOpen: true,
+      warning: ['jadwal', 'rute', 'pemesanan', 'service', 'akses', 'user', 'supir']
+    }
+  },
+  methods: {
+    checkMenu(menu) {
+      if (!this.warning.includes(menu)) {
+        this.isOpen = !this.isOpen
+      } else {
+        this.$swal('Menu belum tersedia!');
+      }
+    }
+  }
+}
+
+</script>
 <template>
-  <div id="app">
-    <nav class="relative bg-white dark:bg-gray-900 fixed w-full z-20 top-0 left-0 border-b border-gray-200 dark:border-gray-600 ">
-      <div class="max-w-screen-xl flex flex-wrap items-center justify-between mx-auto p-4 ">
-        <router-link :to="{ name: 'home'}" href="https://flowbite.com/" class="flex items-center">
-          <img src="img/tj0.jpg" class="h-8 mr-3" alt="Flowbite Logo">
-          <span class="self-center text-2xl font-semibold whitespace-nowrap dark:text-white">BusKoe</span>
-        </router-link>
-        <div class="flex md:order-2">
-          <button type="button"
-            class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 text-center mr-3 md:mr-0 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">Get
-            started</button>
-          <button data-collapse-toggle="navbar-sticky" type="button"
-            class="inline-flex items-center p-2 text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600"
-            aria-controls="navbar-sticky" aria-expanded="false">
-            <span class="sr-only">Open main menu</span>
-            <svg class="w-6 h-6" aria-hidden="true" fill="currentColor" viewBox="0 0 20 20"
-              xmlns="http://www.w3.org/2000/svg">
-              <path fill-rule="evenodd"
-                d="M3 5a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 10a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 15a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1z"
-                clip-rule="evenodd"></path>
-            </svg>
-          </button>
-        </div>
-        <div class="items-center justify-between hidden w-full md:flex md:w-auto md:order-1" id="navbar-sticky">
-          <ul
-            class="flex flex-col p-4 md:p-0 mt-4 font-medium border border-gray-100 rounded-lg bg-gray-50 md:flex-row md:space-x-8 md:mt-0 md:border-0 md:bg-white dark:bg-gray-800 md:dark:bg-gray-900 dark:border-gray-700">
-            <li>
-              <a href="#"
-                class="block py-2 pl-3 pr-4 text-white bg-blue-700 rounded md:bg-transparent md:text-blue-700 md:p-0 md:dark:text-blue-500"
-                aria-current="page">Home</a>
-            </li>
-            <li>
-              <a href="#"
-                class="block py-2 pl-3 pr-4 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:p-0 md:dark:hover:text-blue-500 dark:text-white dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700">About</a>
-            </li>
-            <li>
-              <a href="#"
-              class="block py-2 pl-3 pr-4 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:p-0 md:dark:hover:text-blue-500 dark:text-white dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700">Services</a>
-          </li>
-          <li>
-            <a href="#"
-              class="block py-2 pl-3 pr-4 text-gray-900 rounded hover:bg-gray-100 md:hover:bg-transparent md:hover:text-blue-700 md:p-0 md:dark:hover:text-blue-500 dark:text-white dark:hover:bg-gray-700 dark:hover:text-white md:dark:hover:bg-transparent dark:border-gray-700">Contact</a>
-          </li>
-        </ul>
+  <div class="bg-cyan-700 h-screen w-full px-24">
+    <header class="pt-10 flex justify-items-center">
+      <img src="./assets/logo.png" class="w-10 h-10" />
+      <span class="text-white text-2xl ml-2 font-semibold">Management Travel</span>
+    </header>
+    <div class="mt-20" v-if="isOpen">
+      <h1 class="text-4xl font-semibold text-white mb-10">Startup</h1>
+      <div class="flex gap-2">
+        <RouterLink to="/jadwal" @click="checkMenu('jadwal')" class="w-96 h-44 bg-red-600 flex-row hover:bg-red-700">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/jadwal.png" class="w-28 h-28" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">Jadwal</span>
+            </div>
+            <div v-if="warning.includes('jadwal')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
+        <RouterLink to="/rute" @click="checkMenu('rute')" class="w-96 h-44 bg-sky-400 flex-row hover:bg-sky-500">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/rute.png" class="w-28 h-28" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">Rute</span>
+            </div>
+            <div v-if="warning.includes('rute')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
+        <RouterLink to="/pemesanan" @click="checkMenu('pemesanan')"
+          class="w-44 h-44 bg-amber-500 flex-row hover:bg-amber-600">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/bayar.svg" class="w-28 h-28" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">Pemesanan</span>
+            </div>
+            <div v-if="warning.includes('pemesanan')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
+        <RouterLink to="/service" @click="checkMenu('service')"
+          class="w-96 h-44 bg-green-800 flex-row hover:bg-green-900">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/service.svg" class="w-28 h-28" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">Service</span>
+            </div>
+            <div v-if="warning.includes('service')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
+      </div>
+      <div class="flex gap-2 mt-2">
+        <RouterLink to="/akses" @click="checkMenu('akses')" class="w-44 h-44 bg-lime-500 flex-row hover:bg-lime-600">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/akses.svg" class="w-28 h-28" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">Akses</span>
+            </div>
+            <div v-if="warning.includes('akses')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
+        <RouterLink to="/user" @click="checkMenu('user')" class="w-96 h-44 bg-blue-800 flex-row hover:bg-blue-900">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/user.png" class="w-24 h-24" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">User</span>
+            </div>
+            <div v-if="warning.includes('user')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
+        <RouterLink to="/supir" @click="checkMenu('supir')" class="w-96 h-44 bg-yellow-400 flex-row hover:bg-yellow-500">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/supir.png" class="w-28 h-28" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">User</span>
+            </div>
+            <div v-if="warning.includes('supir')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
+        <RouterLink to="/bus" @click="checkMenu('bus')" class="w-96 h-44 bg-purple-600 flex-row hover:bg-purple-700">
+          <div class="flex w-full h-3/4 justify-center items-center">
+            <img src="./assets/icon/bus.svg" class="w-32 h-32" />
+          </div>
+          <div class="w-full h-1/4 text-left flex flex justify-between px-5">
+            <div>
+              <span class="text-2xl text-white font-normal mt-10">Bus</span>
+            </div>
+            <div v-if="warning.includes('bus')">
+              <img src="./assets/warning.png" class="w-8 h-8" />
+            </div>
+          </div>
+        </RouterLink>
       </div>
     </div>
-  </nav>
-  <router-view />
-</div></template>
+    <div class="mt-5 pt-5 text-white border-t-2 border-white" v-if="!isOpen">
+      <div class="text-blue-300 underline text-lg pb-10 hover:text-blue-500 cursor-pointer" @click="checkMenu('back')">
+        &#x3C; Kembali</div>
+      <RouterView />
+    </div>
+  </div>
+</template>
